HarvestTest now works - inserting a harvest into the db
git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/0.3.0@909 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-common/src/java/xc/mst/dao/BaseDAO.java
@@ -71,11 +71,17 @@ public class BaseDAO {
 	}
 	
 	public void createSchema(String name) {
+		createSchema(name, false);
+	}
+	
+	public void createSchema(String name, boolean dropIfExists) {
+		name = name.toLowerCase().replaceAll(" ", "_");
 		// this is potentially dangerous, but necessary for now
-		if (schemasExists(name)) {
+		if (dropIfExists && schemasExists(name)) {
 			deleteSchema(name);
 		}
-		this.jdbcTemplate.execute("create database "+name+" character set=utf8;");
+		if (!schemasExists(name))
+			this.jdbcTemplate.execute("create database "+name+" character set=utf8;");
 	}
 	
 	public void deleteSchema(String name) {
